+++ src/app/admin/gallery/page.tsx
@@ -1,25 +1,7 @@
-import { GalleryItem } from "@prisma/client";
-import { columns } from "./columns";
+import GalleryDatabase from "@/components/admin/event/EventDatabase";
 
-import { DataTable } from "./data-table";
-
-async function getEventData(): Promise<GalleryItem[]> {
-	const events = await fetch("http://ruvsa.vercel.app/api/gallery", {
-		next: { revalidate: 1 },
-	});
-
-	return events.json();
-}
-
-export default async function GalleryDatabase() {
-	const data = await getEventData();
-
-	return (
-		<div>
-			<h1>Edit Gallery Database</h1>
-			<div className="container mx-auto py-10">
-				<DataTable columns={columns} data={data} />
-			</div>
-		</div>
-	);
+export default function AdminBoard() {
+   <div>
+      <GalleryDatabase />
+   </div>;
 }
